Error page rendering overhaul.  Multi-path TemplateLoader
This change merges rendering of error pages in to the TemplateLoader
mechanism.  As part of this, the TemplateLoader has been changed to
look in multiple directories, in priority order.  This will allow
template overriding (e.g. the app provides its own error page template,
if desired).  Lastly, it allows the TemplateLoader to be used even when
not all templates are correct.  (e.g. in the case of error pages)