Added Accept-Language HTTP request header parsing
For each request the Accept-Language HTTP header is now parsed. The wrapping
Revel contains a "AcceptLanguages" field that contains the sorted slice of
all Accept-Language values if there were any. The first language in the slice
is the most qualified.

If the Accept-Language HTTP header was not set or contained no value, the
value of the AcceptLanguages field is set to nil.