Benchmark for the Revel request flow overhead
This is meant to measure the Revel overhead (routing, reflective
controller invocation, parameter binding, Render() call magic, template
rendering).  It does *not* make any network connections, so it does not
measure the performance of the actual server.

Comes out to 200us per on my macbook, or 1ms including the 10 static
requests.