Convert Filter type to a func
- This halves the stack trace -- every interface call adds 2 stack
frames.  Since the filters are all nested, it got pretty overwhelming.
- Reduces the boilerplate for all the built-in filters (net -50 lines).
 I think that package level variables are always fine (in fact
preferred) for the filter, so the struct type gives no benefit.
- Reduces the ways to get into OnAppStart back to 1.

Downsides:
- Now the filters do not have a name that can be referenced in error
messages , etc.
- It is common for a filter to have some initialization, so it's a
little too bad we can't syntactically sugar that a little more than the
usual OnAppStart declaration.

Also, remove the FilterChain type as it was not adding anything.