Add "X-HTTP-Method-Override" to router
This will add X-HTTP-Method-Override as a method override to the router.
The request method is replaced, so subsequent filters can test on the
intended method.

This should make it safe for CSRF testing, since the CSRF filter will
see the intended method and not the actual used one.

Test added. I found no way to declare the headers inline, so it has to
be added in code.