Add i18n templates.
All templates can now have an optional language suffix. The loader will attempt to load a template with the language specified language as an extension to the specified name. For instance "index.html" with "en" as language will attempt to load "index.html.en" first, and fall back to "template.html" if the file cannot be found.

Internationalized templates can also be dynamically included in the templates. To do this, use a the template function {{ i18template . "included.html" }}

This function will also attempt to load "included.html.en" if the language is English, and will fall back to including "included.html" if the language cannot be found.